Move test_units into test_weights_measures
